+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 - Do not log warning for component first emission in `melody-streams` in production [#128](https://github.com/trivago/melody/pull/128)
 - Remove prefetch for async components [#130](https://github.com/trivago/melody/pull/130)
+- Fixed `~` operator must convert value to string [#123](https://github.com/trivago/melody/pull/123)
 
 ## 1.2.0